# [POC] Optimize Logs SDK batch processing