--- a/sdk/log/batch.go
+++ b/sdk/log/batch.go
@@ -75,7 +75,7 @@ type BatchProcessor struct {
 	exporter *bufferExporter
 
 	// q is the active queue of records that have not yet been exported.
-	q *queue
+	q *cqueue
 	// batchSize is the minimum number of records needed before an export is
 	// triggered (unless the interval expires).
 	batchSize int
@@ -122,7 +122,7 @@ func NewBatchProcessor(exporter Exporter, opts ...BatchProcessorOption) *BatchPr
 	b := &BatchProcessor{
 		exporter: newBufferExporter(exporter, cfg.expBufferSize.Value),
 
-		q:           newQueue(cfg.maxQSize.Value),
+		q:           newCQueue(cfg.maxQSize.Value),
 		batchSize:   cfg.expMaxBatchSize.Value,
 		pollTrigger: make(chan struct{}, 1),
 		pollKill:    make(chan struct{}),
@@ -138,7 +138,7 @@ func (b *BatchProcessor) poll(interval time.Duration) (done chan struct{}) {
 
 	ticker := time.NewTicker(interval)
 	// TODO: investigate using a sync.Pool instead of cloning.
-	buf := make([]Record, b.batchSize)
+	buf := make([]*Record, b.batchSize)
 	go func() {
 		defer close(done)
 		defer ticker.Stop()
@@ -156,7 +156,7 @@ func (b *BatchProcessor) poll(interval time.Duration) (done chan struct{}) {
 				global.Warn("dropped log records", "dropped", d)
 			}
 
-			qLen := b.q.TryDequeue(buf, func(r []Record) bool {
+			qLen := b.q.TryDequeue(buf, func(r []*Record) bool {
 				ok := b.exporter.EnqueueExport(r)
 				if ok {
 					buf = slices.Clone(buf)
@@ -184,7 +184,8 @@ func (b *BatchProcessor) OnEmit(_ context.Context, r *Record) error {
 	}
 	// The record is cloned so that changes done by subsequent processors
 	// are not going to lead to a data race.
-	if n := b.q.Enqueue(r.Clone()); n >= b.batchSize {
+	rc := r.Clone()
+	if n := b.q.Enqueue(&rc); n >= b.batchSize {
 		select {
 		case b.pollTrigger <- struct{}{}:
 		default:
@@ -229,10 +230,10 @@ func (b *BatchProcessor) ForceFlush(ctx context.Context) error {
 		return nil
 	}
 
-	buf := make([]Record, b.q.cap)
+	buf := make([]*Record, b.q.cap)
 	notFlushed := func() bool {
 		var flushed bool
-		_ = b.q.TryDequeue(buf, func(r []Record) bool {
+		_ = b.q.TryDequeue(buf, func(r []*Record) bool {
 			flushed = b.exporter.EnqueueExport(r)
 			return flushed
 		})
@@ -283,7 +284,7 @@ func (q *queue) Dropped() uint64 {
 //
 // If enqueueing r will exceed the capacity of q, the oldest Record held in q
 // will be dropped and r retained.
-func (q *queue) Enqueue(r Record) int {
+func (q *queue) Enqueue(r *Record) int {
 	q.Lock()
 	defer q.Unlock()
 
@@ -308,7 +309,7 @@ func (q *queue) Enqueue(r Record) int {
 //
 // When write is called the lock of q is held. The write function must not call
 // other methods of this q that acquire the lock.
-func (q *queue) TryDequeue(buf []Record, write func([]Record) bool) int {
+func (q *queue) TryDequeue(buf []*Record, write func([]*Record) bool) int {
 	q.Lock()
 	defer q.Unlock()
 
@@ -330,11 +331,11 @@ func (q *queue) TryDequeue(buf []Record, write func([]Record) bool) int {
 
 // Flush returns all the Records held in the queue and resets it to be
 // empty.
-func (q *queue) Flush() []Record {
+func (q *queue) Flush() []*Record {
 	q.Lock()
 	defer q.Unlock()
 
-	out := make([]Record, q.len)
+	out := make([]*Record, q.len)
 	for i := range out {
 		out[i] = q.read.Value
 		q.read = q.read.Next()