Proof of concept fix for: https://github.com/open-telemetry/opentelemetry-go/issues/6382

Resolutions steps:

- refactored to use of pointers to log records instead of passing by values (~1.8x improvement)
- reimplement queue (ring based) as cyclic array queue (~16x improvement)

### Benchmarks https://github.com/pellared/spanevents-vs-logs

#### Base (OTLP)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │                 logs.out                 │
         │     sec/op     │     sec/op      vs base                  │
/OTLP-24      43.58µ ± 6%   2977.49µ ± 10%  +6731.84% (p=0.000 n=10)

         │ spanevents.out │               logs.out                │
         │      B/op      │     B/op       vs base                │
/OTLP-24     99.47Ki ± 5%   193.25Ki ± 1%  +94.29% (p=0.000 n=10)

         │ spanevents.out │               logs.out               │
         │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24      710.0 ± 12%   1861.0 ± 2%  +162.11% (p=0.000 n=10)
```
#### Pointer (OTLP)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │                 logs.out                 │
         │     sec/op     │     sec/op      vs base                  │
/OTLP-24      43.58µ ± 6%   1617.18µ ± 25%  +3610.63% (p=0.000 n=10)

         │ spanevents.out │                logs.out                │
         │      B/op      │      B/op       vs base                │
/OTLP-24     99.47Ki ± 5%   148.22Ki ± 28%  +49.01% (p=0.002 n=10)

         │ spanevents.out │               logs.out               │
         │   allocs/op    │  allocs/op    vs base                │
/OTLP-24      710.0 ± 12%   1196.5 ± 56%  +68.52% (p=0.023 n=10)
```

#### Pointer Array Queue (OTLP)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
         │ spanevents.out │               logs.out                │
         │     sec/op     │    sec/op     vs base                 │
/OTLP-24      42.44µ ± 7%   178.70µ ± 4%  +321.08% (p=0.000 n=10)

         │ spanevents.out │            logs.out            │
         │      B/op      │     B/op      vs base          │
/OTLP-24     99.34Ki ± 3%   98.13Ki ± 3%  ~ (p=0.280 n=10)

         │ spanevents.out │              logs.out               │
         │   allocs/op    │  allocs/op   vs base                │
/OTLP-24       707.0 ± 7%   400.0 ± 11%  -43.42% (p=0.000 n=10)
```

--------------------------------------------------------------------------------------------

#### Base (All)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │ spanevents.out │                 logs.out                 │
           │     sec/op     │     sec/op      vs base                  │
/OTLP-24        46.27µ ± 6%   3180.09µ ± 11%  +6772.15% (p=0.000 n=10)
/STDOUT-24      41.97µ ± 2%    638.88µ ±  1%  +1422.32% (p=0.000 n=10)
/NOOP-24       14.268µ ± 2%     4.496µ ±  1%    -68.49% (p=0.000 n=10)
geomean         30.26µ          209.0µ         +590.81%

           │ spanevents.out │                logs.out                │
           │      B/op      │     B/op       vs base                 │
/OTLP-24       100.5Ki ± 1%    191.4Ki ± 2%   +90.41% (p=0.000 n=10)
/STDOUT-24     90.94Ki ± 0%   274.93Ki ± 0%  +202.33% (p=0.000 n=10)
/NOOP-24       29840.0 ± 0%      240.0 ± 0%   -99.20% (p=0.000 n=10)
geomean        64.34Ki         23.10Ki        -64.09%

           │ spanevents.out │               logs.out               │
           │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24         729.5 ± 3%   1842.0 ± 2%  +152.50% (p=0.000 n=10)
/STDOUT-24       644.5 ± 1%   4625.5 ± 0%  +617.69% (p=0.000 n=10)
/NOOP-24       303.000 ± 0%    3.000 ± 0%   -99.01% (p=0.000 n=10)
geomean          522.3         294.6        -43.60%
```
#### Pointer Array Queue (All)
```
go tool benchstat spanevents.out logs.out
goos: linux
goarch: amd64
pkg: github.com/pellared/spanevents-vs-logs
cpu: AMD Ryzen 9 7900X 12-Core Processor
           │ spanevents.out │               logs.out                │
           │     sec/op     │    sec/op     vs base                 │
/OTLP-24        44.22µ ± 3%   187.59µ ± 4%  +324.19% (p=0.000 n=10)
/STDOUT-24      40.65µ ± 3%   186.79µ ± 4%  +359.53% (p=0.000 n=10)
/NOOP-24       13.825µ ± 3%    4.346µ ± 1%   -68.57% (p=0.000 n=10)
geomean         29.18µ         53.40µ        +82.99%

           │ spanevents.out │               logs.out                │
           │      B/op      │     B/op       vs base                │
/OTLP-24       99.72Ki ± 7%    97.90Ki ± 1%   -1.82% (p=0.000 n=10)
/STDOUT-24     91.00Ki ± 0%   145.70Ki ± 1%  +60.12% (p=0.000 n=10)
/NOOP-24       29840.0 ± 0%      240.0 ± 0%  -99.20% (p=0.000 n=10)
geomean        64.19Ki         14.95Ki       -76.70%

           │ spanevents.out │               logs.out               │
           │   allocs/op    │  allocs/op   vs base                 │
/OTLP-24        714.5 ± 19%    396.5 ± 3%   -44.51% (p=0.000 n=10)
/STDOUT-24      647.5 ±  1%   1581.0 ± 2%  +144.17% (p=0.000 n=10)
/NOOP-24      303.000 ±  0%    3.000 ± 0%   -99.01% (p=0.000 n=10)
geomean         519.5          123.4        -76.24%
```

### Benchmark (this repository)

* base.txt - current implementation
* ptr.txt - pointer refactor
* fullfix.txt - pointer refactor + cyclic array queue

```sh
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/log
cpu: AMD Ryzen 9 7900X 12-Core Processor
                                 │  base.txt   │               ptr.txt                │             fullfix.txt             │
                                 │   sec/op    │    sec/op     vs base                │   sec/op     vs base                │
Processor/Simple-24                337.6n ± 2%   368.9n ± 12%        ~ (p=0.105 n=10)   324.6n ± 3%   -3.87% (p=0.001 n=10)
Processor/Batch-24                 679.0n ± 5%   552.6n ±  2%  -18.62% (p=0.000 n=10)   481.2n ± 2%  -29.13% (p=0.000 n=10)
Processor/SetTimestampSimple-24    361.0n ± 1%   330.7n ±  1%   -8.41% (p=0.000 n=10)   327.7n ± 1%   -9.24% (p=0.000 n=10)
Processor/SetTimestampBatch-24     672.5n ± 6%   550.6n ±  3%  -18.13% (p=0.000 n=10)   479.4n ± 1%  -28.71% (p=0.000 n=10)
Processor/AddAttributesSimple-24   392.7n ± 2%   350.8n ±  1%  -10.65% (p=0.000 n=10)   353.4n ± 1%  -10.01% (p=0.000 n=10)
Processor/AddAttributesBatch-24    767.1n ± 4%   585.9n ±  4%  -23.63% (p=0.000 n=10)   509.1n ± 1%  -33.64% (p=0.000 n=10)
Processor/SetAttributesSimple-24   375.9n ± 4%   339.6n ±  1%   -9.66% (p=0.000 n=10)   334.9n ± 1%  -10.92% (p=0.000 n=10)
Processor/SetAttributesBatch-24    704.4n ± 4%   573.1n ±  4%  -18.63% (p=0.000 n=10)   491.6n ± 1%  -30.21% (p=0.000 n=10)
geomean                            508.0n        443.0n        -12.79%                  405.2n       -20.25%

                                 │  base.txt   │               ptr.txt                │             fullfix.txt              │
                                 │    B/op     │    B/op     vs base                  │    B/op     vs base                  │
Processor/Simple-24                 450.0 ± 0%   450.0 ± 0%        ~ (p=0.474 n=10)     449.5 ± 0%   -0.11% (p=0.033 n=10)
Processor/Batch-24                 1173.5 ± 1%   916.0 ± 0%  -21.94% (p=0.000 n=10)     915.0 ± 0%  -22.03% (p=0.000 n=10)
Processor/SetTimestampSimple-24     450.0 ± 0%   450.0 ± 0%        ~ (p=1.000 n=10) ¹   450.0 ± 0%        ~ (p=1.000 n=10) ¹
Processor/SetTimestampBatch-24     1175.0 ± 1%   916.0 ± 0%  -22.04% (p=0.000 n=10)     915.0 ± 0%  -22.13% (p=0.000 n=10)
Processor/AddAttributesSimple-24    450.0 ± 0%   450.0 ± 0%        ~ (p=1.000 n=10) ¹   450.0 ± 0%        ~ (p=1.000 n=10) ¹
Processor/AddAttributesBatch-24    1194.5 ± 1%   916.0 ± 0%  -23.32% (p=0.000 n=10)     915.0 ± 0%  -23.40% (p=0.000 n=10)
Processor/SetAttributesSimple-24    450.0 ± 0%   450.0 ± 0%        ~ (p=1.000 n=10) ¹   450.0 ± 0%        ~ (p=1.000 n=10) ¹
Processor/SetAttributesBatch-24    1189.5 ± 2%   916.0 ± 0%  -22.99% (p=0.000 n=10)     915.0 ± 0%  -23.08% (p=0.000 n=10)
geomean                             729.7        642.0       -12.01%                    641.6       -12.07%
¹ all samples are equal

                                 │  base.txt  │                ptr.txt                │              fullfix.txt              │
                                 │ allocs/op  │ allocs/op   vs base                   │ allocs/op   vs base                   │
Processor/Simple-24                1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹   1.000 ± 0%         ~ (p=1.000 n=10) ¹
Processor/Batch-24                 1.000 ± 0%   2.000 ± 0%  +100.00% (p=0.000 n=10)     2.000 ± 0%  +100.00% (p=0.000 n=10)
Processor/SetTimestampSimple-24    1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹   1.000 ± 0%         ~ (p=1.000 n=10) ¹
Processor/SetTimestampBatch-24     1.000 ± 0%   2.000 ± 0%  +100.00% (p=0.000 n=10)     2.000 ± 0%  +100.00% (p=0.000 n=10)
Processor/AddAttributesSimple-24   1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹   1.000 ± 0%         ~ (p=1.000 n=10) ¹
Processor/AddAttributesBatch-24    1.000 ± 0%   2.000 ± 0%  +100.00% (p=0.000 n=10)     2.000 ± 0%  +100.00% (p=0.000 n=10)
Processor/SetAttributesSimple-24   1.000 ± 0%   1.000 ± 0%         ~ (p=1.000 n=10) ¹   1.000 ± 0%         ~ (p=1.000 n=10) ¹
Processor/SetAttributesBatch-24    1.000 ± 0%   2.000 ± 0%  +100.00% (p=0.000 n=10)     2.000 ± 0%  +100.00% (p=0.000 n=10)
geomean                            1.000        1.414        +41.42%                    1.414        +41.42%
```