@pellared 2025-03-27
Can you please paste the benchmarks from here: https://github.com/open-telemetry/opentelemetry-go/pull/6493#issuecomment-2748921789 in the PR description?